Rename Marker to StakeholderLayer, use forwardRef to pass through ref
